Adding FluentUIDemoToggle.
