+++ ios/FluentUI.Demo/FluentUI.Demo/Demos/AvatarDemoController_SwiftUI.swift
@@ -37,16 +37,6 @@ struct AvatarDemoView: View {
     @State var size: MSFAvatarSize = .xxlarge
     @State var style: MSFAvatarStyle = .default
 
-    let switchToggleStyle: SwitchToggleStyle = {
-        if #available(iOS 14.0, *) {
-            // Workaround for SwiftUI Toggles that switch to green tint color
-            // when they're toggled for the first time
-            return SwitchToggleStyle(tint: Color(Colors.communicationBlue))
-        } else {
-            return SwitchToggleStyle()
-        }
-    }()
-
     public var body: some View {
         VStack {
             AvatarView(style: style,
@@ -84,14 +74,10 @@ struct AvatarDemoView: View {
                             .disableAutocorrection(true)
                             .textFieldStyle(RoundedBorderTextFieldStyle())
 
-                        Toggle("Set image", isOn: $showImage)
-                            .toggleStyle(switchToggleStyle)
-                        Toggle("Set alternate background", isOn: $useAlternateBackground)
-                            .toggleStyle(switchToggleStyle)
-                        Toggle("Transparency", isOn: $isTransparent)
-                            .toggleStyle(switchToggleStyle)
-                        Toggle("iPad Pointer interaction", isOn: $hasPointerInteraction)
-                            .toggleStyle(switchToggleStyle)
+                        FluentUIDemoToggle(titleKey: "Set image", isOn: $showImage)
+                        FluentUIDemoToggle(titleKey: "Set alternate background", isOn: $useAlternateBackground)
+                        FluentUIDemoToggle(titleKey: "Transparency", isOn: $isTransparent)
+                        FluentUIDemoToggle(titleKey: "iPad Pointer interaction", isOn: $hasPointerInteraction)
                     }
 
                     Group {
@@ -101,12 +87,9 @@ struct AvatarDemoView: View {
                                 .font(.title)
                             Divider()
                         }
-                        Toggle("Ring visible", isOn: $isRingVisible)
-                            .toggleStyle(switchToggleStyle)
-                        Toggle("Ring inner gap", isOn: $hasRingInnerGap)
-                            .toggleStyle(switchToggleStyle)
-                        Toggle("Set image based ring color", isOn: $showImageBasedRingColor)
-                            .toggleStyle(switchToggleStyle)
+                        FluentUIDemoToggle(titleKey: "Ring visible", isOn: $isRingVisible)
+                        FluentUIDemoToggle(titleKey: "Ring inner gap", isOn: $hasRingInnerGap)
+                        FluentUIDemoToggle(titleKey: "Set image based ring color", isOn: $showImageBasedRingColor)
                     }
 
                     Group {
@@ -130,8 +113,7 @@ struct AvatarDemoView: View {
                         .labelsHidden()
                         .frame(maxWidth: .infinity, alignment: .leading)
 
-                        Toggle("Out of office", isOn: $isOutOfOffice)
-                            .toggleStyle(switchToggleStyle)
+                        FluentUIDemoToggle(titleKey: "Out of office", isOn: $isOutOfOffice)
                     }
 
                     Group {
